refactor(sidebar): extract header into SidebarHeader component
Extracted the header section, `UnicornHeadIcon`, `HeaderButton`, theme configurations,
and related logic out of `Sidebar.tsx` into a new `SidebarHeader.tsx` component.
This drastically reduces the complexity and line count of the "Long Function" `Sidebar` component.
State handlers required by the header were maintained via `useGraphStore` where appropriate
or passed down via props. All tests and linting passed without regressions.
